AOC 2025 Day 4 Intcode Solver - Back to assertion failures
I fixed the counting error that caused it to mark everything as eligible for removal.

That revealed an infinite loop when a roll is *not* eligible for removal (it did not advance through the linked list, so it would scan the roll forever).

I fixed _that_ bug, and now it's back to trying to remove the roll from a cell that doesn't have one.