AOC 2025 Day 8 Intcode - WIP
Gotten a lot farther this time.

It looks like it's able to process the pairs heap without errors now, which is nice.

For some reason it doesn't stop when the heap is empty, though, causing an assertion failure.

The part 1 answer seems to be wrong, so something's probably wrong with how I'm building the circuit data.