AOC 2025 Day 9 Intcode draw_perimeter WIP
I was setting up the initial variables wrong, so fixing that makes things better.

However, it fails on the third or fourth line.

Looking more closely, it appears that Y-compression is corrupting the tile coordinates somehow; after X-compression, everything is still proper, but after Y-compression, one of the perimeter lines is no longer orthogonal. Odd.