AOC 2025 Day 9 Intcode Fixed non-compressed mode calculations
The final grid now looks almost correct:

.......EEE
.......#.E
EEEEEEE#.E
#........E
######...E
.....#...E
.....####E

If I flipped the Y-axis again, those Es would all be in exactly the right places.
But I can't flip the Y-axis again, because that would make the grid be out of sync with the tile coordinates.
I had did the initial flip to try and make the Es be in the right place.  What am I doing wrong?