AOC 2025 Day 9 Intcode - SO CLOSE
I fixed the unnecessary (and invalid) correction that caused it to only try the first coordinate against all the others.

Now it gets the right answer for the example for part 1, but not for part 2.  For some reason, the correct answer is not accepted.
I'll have to figure out why...

(7, 4) - (0, 2)

```
.....####E
.....#...E
######...E
#........E
EEEEEEE#.E
.......#.E
.......EEE
```